BUG: integer

+++ pysatNASA/instruments/methods/omni.py
@@ -113,7 +113,7 @@ def calculate_imf_steadiness(inst, steady_window=15, min_window_frac=0.75,
     sample_rate = int(rates[inst.tag])
     max_wnum = np.floor(steady_window / sample_rate)
     if max_wnum != steady_window / sample_rate:
-        steady_window = max_wnum * sample_rate
+        steady_window = int(max_wnum * sample_rate)
         pysat.logger.warning(" ".join(("sample rate is not a factor of the",
                                        "statistical window")))
         pysat.logger.warning(" ".join(("new statistical window is",
